Split AnswerUtils into local usages
