+++ graql/reasoner/state/VariableComparisonState.java
@@ -20,11 +20,11 @@
 package grakn.core.graql.reasoner.state;
 
 import com.google.common.collect.Iterators;
-import grakn.core.concept.answer.AnswerUtil;
 import grakn.core.concept.answer.ConceptMap;
 import grakn.core.graql.reasoner.atom.predicate.VariablePredicate;
 import grakn.core.graql.reasoner.query.ReasonerAtomicQuery;
 import grakn.core.graql.reasoner.query.ReasonerQueryImpl;
+import grakn.core.graql.reasoner.utils.AnswerUtil;
 import grakn.core.kb.graql.reasoner.unifier.Unifier;
 
 import java.util.Iterator;
