+++ graql/reasoner/utils/AnswerUtil.java
@@ -17,60 +17,24 @@
  *
  */
 
-package grakn.core.concept.answer;
+package grakn.core.graql.reasoner.utils;
 
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Sets;
+import grakn.core.concept.answer.ConceptMap;
 import grakn.core.concept.util.ConceptUtils;
 import grakn.core.kb.concept.api.Concept;
 import grakn.core.kb.concept.api.SchemaConcept;
 import graql.lang.Graql;
 import graql.lang.pattern.Pattern;
-import graql.lang.query.builder.Filterable;
 import graql.lang.statement.Variable;
 
-import java.util.Comparator;
 import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-/**
- * Helper Class for manipulating different types of answers
- * Used from within reasoner, query execution, and tests -- code regions generating answers.
- * Located with `answer` implementations to be accessible from all of these places, and because
- * it makes sense to group answers and answer manipulation together
- */
 public class AnswerUtil {
-
-    @SuppressWarnings("unchecked") // All attribute values are comparable data types
-    public static Stream<ConceptMap> filter(Filterable query, Stream<ConceptMap> answers) {
-        if (query.sort().isPresent()) {
-            Variable var = query.sort().get().var();
-            Comparator<ConceptMap> comparator = (map1, map2) -> {
-                Object val1 = map1.get(var).asAttribute().value();
-                Object val2 = map2.get(var).asAttribute().value();
-
-                if (val1 instanceof String) {
-                    return ((String) val1).compareToIgnoreCase((String) val2);
-                } else {
-                    return ((Comparable<? super Comparable>) val1).compareTo((Comparable<? super Comparable>) val2);
-                }
-            };
-            comparator = (query.sort().get().order() == Graql.Token.Order.DESC) ? comparator.reversed() : comparator;
-            answers = answers.sorted(comparator);
-        }
-        if (query.offset().isPresent()) {
-            answers = answers.skip(query.offset().get());
-        }
-        if (query.limit().isPresent()) {
-            answers = answers.limit(query.limit().get());
-        }
-
-        return answers;
-    }
-
-
     /**
      * Performs a natural join (⋈) between answers with the resultant answer containing the explanation of the left operand.
      * If the answers have an empty set of common variables, the join corresponds to a trivial entry set union.
