+++ graql/test/reasoner/BUILD
@@ -28,6 +28,7 @@ java_test(
         "//dependencies/maven/artifacts/com/google/guava",
         "//dependencies/maven/artifacts/org/mockito:mockito-core",
         "//kb/concept/api",
+        "//graql",
         "@graknlabs_graql//java:graql",
     ],
 )
